+++ terraform/test-fixtures/apply-destroy-provisioner/main.tf
@@ -1,0 +1,5 @@
+resource "aws_instance" "foo" {
+    id = "foo"
+
+    provisioner "shell" {}
+}