+++ CHANGELOG.md
@@ -11,6 +11,7 @@ BUG FIXES:
       marked as tainted. [GH-434]
   * core: Set types are validated to be sets. [GH-413]
   * core: Fix crash case when destroying with tainted resources. [GH-412]
+  * core: Don't execute provisioners in some cases on destroy.
   * providers/aws: Refresh of launch configs and autoscale groups load
       the correct data and don't incorrectly recreate themselves. [GH-425]
   * providers/aws: Fix case where ELB would incorrectly plan to modify